requirements.txt
jax
jaxlib
chex
dm-haiku